--- Assets/Art/Models.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 90379946f80e7a347b1530959dda1265
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
