+++ Assets/Scenes/Tony's Work/Tony's/LightingData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b6f75abd029349d468da79b02ec8e085
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 25800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
