+++ Assets/Scenes/Tony's Work/Tony's/Lightmap-0_comp_dir.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:40f58f8f1b4c704dafdfb4c6f41abdfdecbff33fee780af8f303b887b5a2ed7b
+size 11857
