+++ Assets/Scenes/Tony's Work/Tony's/Lightmap-1_comp_dir.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:10a5696276aa4aa58d9b56a97ba0ace07d19dfae53de6accc55187018201c57a
+size 11461
