+++ Assets/Scenes/Tony's Work/Tony's/Lightmap-2_comp_dir.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6eb8348f7ff41f09fa147dd6258be1395ce18e0ac6af12589a738ea077b6f96e
+size 7969
