--- Assets/Scenes/Tony's.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5523d779d84343c47a18d8df755ee30c
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
