--- Assets/Scripts/InputManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e3ccb633bb537934d8e344e5222b5d7f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
