+++ Assets/Scripts/SimonSays.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: edf8ba9a94ee4ba49b244ec384a9b71f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
