+++ Assets/Scripts/SimonSays/SimonImageDisplay.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class SimonImageDisplay : MonoBehaviour {
+
+
+
+    public void DisplayImageSequence(Queue<Colors> colors)
+    {
+        if (colors.Enqueue()
+        this.GetComponent<Image>().color = colors.Dequeue();
+    }
+
+        
+
+}
