+++ Assets/Scripts/SimonSays/SimonManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a489fbb513c87a54ab6aae72f8ff0a95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
