--- Assets/Scripts/SimonSaysRandomizer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a99535e1073862a4f9bf0d3495abdab1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
