+++ ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 2018.1.6f1
+m_EditorVersion: 2018.1.9f1
